llm-processor/requirements.txt
fastapi==0.110.0
uvicorn==0.30.1
langchain==0.2.3
langgraph==0.2.28
langchain-core==0.2.40
langchain-community==0.2.0
langchain-openai==0.1.23
langchain-google-genai==1.0.10
langsmith==0.1.112
pinecone-client==3.2.2
httpx==0.27.0
pydantic==2.7.4
python-dotenv==1.0.1
PyYAML==6.0.1
firecrawl-py==4.5.0
